// src/app/pages/stock/pages/tools-manager-page/interfaces/tool.interface.ts
export interface ITool {
  _id?: string;
  Nombre: string;
  Marca: string;
  imgRef: IToolImgRef;
  Stock: {
    Fecha_Compra: string;
    Cantidad: number;
    Estado: string;
    Ubicacion: string;
  }[];
}

export interface IToolImgRef {
  asset_id: string;
  public_id: string;
  version: number;
  version_id: string;
  signature: string;
  width: string;
  height: string;
  format: string;
  resource_type: string;
  created_at: string;
  tags: any[];
  bytes: number;
  type: string;
  etag: string;
  placeholder: false;
  url: string;
  secure_url: string;
  folder: string;
  original_filename: string;
  api_key: string;
}